added new file structure for whole project
