+++ src/components/Layout/Footer/Footer.jsx
@@ -1,4 +1,4 @@
-import logo from "../images/logo.svg";
+import logo from "../../../images/logo.svg";
 
 export default function Footer() {
     return (
