--- src/components/Logo.jsx
@@ -1,6 +1,0 @@
-
-export default function logo() {
-    return (
-        <p>ESSENTIALS</p>
-        )
-}
