+++ src/components/cards/CustomerReview/CustomerReview.jsx
@@ -1,0 +1,9 @@
+export default function CustomerReview ({ review }) {
+    return (
+        <div className="my-5 py-2">
+            <p className="font-semibold text-xl">{review.username}</p>
+            <p>{review.description}</p>
+            <p className="font-semibold">Rating: {review.rating}</p>
+        </div>
+    );
+}
