+++ src/routes/Checkout.jsx
@@ -1,10 +1,12 @@
-import CheckoutTest from "../components/Checkout"
+import Cart from "../components/Cart/Cart"
+import CheckoutBtn from "../components/buttons/CheckoutBtn/ChekoutBtn"
 
 export default function Checkout() {
     return (
        <>
        <div className="h-16"></div>
-       <CheckoutTest/>
+       <Cart/>
+       <CheckoutBtn />
        </>
     )
 }
